adding json logic, still need to implement json

+++ cluster_data_ingestion/gene_files_writer.py
@@ -17,7 +17,7 @@ class GeneFilesWriter:
         else:
             print("Non Existent kingdom kwarg, write 'plants' or 'fungi'")
             raise Exception
-        self.genbank_files = directory_data.GENBANK_3_1
+        self.genbank_files = [directory_data.GENBANK_3_1, directory_data.GENBANK_3_0, directory_data.GENBANK_2_0]
         self._check_create_dir()
 
     def _check_create_dir(self):
@@ -31,28 +31,32 @@ class GeneFilesWriter:
     def _iterate_and_create_files(self, cluster_list):
         for cluster_name in cluster_list:
             cluster_file_name = f'{cluster_name}.gbk'
-            file_path = os.path.join(self.genbank_files, cluster_file_name)
-
-            for seq_record in SeqIO.parse(file_path, IODataTypes.GENBANK):
-                features = seq_record.features
-                df = pd.DataFrame()
-                for feature in features:
-                    gene_row = dict()
-                    gene_row['feature_type'] = feature.type
-                    gene_row['start'] = feature.location.start
-                    gene_row['end'] = feature.location.end
-                    gene_row['gene'] = self.get_value(feature.qualifiers.get('gene'))
-                    gene_row['product'] = self.get_value(feature.qualifiers.get('product'))
-                    gene_row['transcript_id'] = self.get_value(feature.qualifiers.get('transcript_id'))
-                    gene_row['db_xref'] = self.get_value(feature.qualifiers.get('db_xref'))
-                    gene_row['protein_id'] = self.get_value(feature.qualifiers.get('protein_id'))
-                    gene_row['gene_kind'] = self.get_value(feature.qualifiers.get('gene_kind'))
-                    gene_row['note'] = self.get_value(feature.qualifiers.get('note'))
-                    df2 = pd.Series(gene_row).to_frame().transpose()
-                    df = pd.concat([df, df2])
-                filepath = os.path.join(self.write_directory, f'{cluster_name}.xlsx')
-                df.reset_index(drop=True, inplace=True)
-                df.to_excel(filepath, index=True)
+            for directory in self.genbank_files:
+                file_path = os.path.join(directory, cluster_file_name)
+                if not os.path.exists(file_path):
+                    continue
+                for seq_record in SeqIO.parse(file_path, IODataTypes.GENBANK):
+                    features = seq_record.features
+                    df = pd.DataFrame()
+                    for feature in features:
+                        gene_row = dict()
+                        gene_row['feature_type'] = feature.type
+                        gene_row['start'] = feature.location.start
+                        gene_row['end'] = feature.location.end
+                        gene_row['gene'] = self.get_value(feature.qualifiers.get('gene'))
+                        gene_row['product'] = self.get_value(feature.qualifiers.get('product'))
+                        gene_row['transcript_id'] = self.get_value(feature.qualifiers.get('transcript_id'))
+                        gene_row['db_xref'] = self.get_value(feature.qualifiers.get('db_xref'))
+                        gene_row['protein_id'] = self.get_value(feature.qualifiers.get('protein_id'))
+                        gene_row['gene_kind'] = self.get_value(feature.qualifiers.get('gene_kind'))
+                        gene_row['note'] = self.get_value(feature.qualifiers.get('note'))
+                        df2 = pd.Series(gene_row).to_frame().transpose()
+                        df = pd.concat([df, df2])
+                    filepath = os.path.join(self.write_directory, f'{cluster_name}.xlsx')
+                    df.reset_index(drop=True, inplace=True)
+                    df.to_excel(filepath, index=True)
+                print(f"Created {cluster_name} genbank file")
+                break
 
     @staticmethod
     def get_value(lst, index=0):
